+++ src/main/java/D/Entities/Prof.java
@@ -3,7 +3,7 @@ package D.Entities;
 public class Prof extends Person{
 private String type;
 
-    public Prof(Long id, String firstname, String lastname, String username, String password, String type) {
+    public Prof(Integer id, String firstname, String lastname, String username, String password, String type) {
         super(id, firstname, lastname, username, password);
         this.type = type;
     }
