+++ src/main/java/D/Repository/EmployeeRepository.java
@@ -2,5 +2,5 @@ package D.Repository;
 
 import D.Entities.Employee;
 
-public interface EmployeeRepository extends BaseRepository<Employee,Long> {
+public interface EmployeeRepository extends BaseRepository<Employee,Integer> {
 }
