+++ src/main/java/D/Repository/StudentRepository.java
@@ -2,5 +2,5 @@ package D.Repository;
 
 import D.Entities.Student;
 
-public interface StudentRepository extends BaseRepository<Student,Long> {
+public interface StudentRepository extends BaseRepository<Student,Integer> {
 }
